RF/BF: EOL mode now controlled by user preferences. Now, Coder and CodeBoxes have their EOL set on creation by user preferences EOL is converted on compile When pasting into Coder, the EOL of the doc is converted to user prefs. The same occurs with the code component. Paste and getEOL functions are not part of the BaseCode class Allow users to control viewing EOLs from prefs in Coder - not ideal, but otherwise it is not an option - can change if not wanted.
